Master of Uno Nerf
--Removed Polymorphine from Master of Transmuation material changing
--Removed Master of Transmutation from Fungicaves pool
++Added Master of Transmutation to Lukki Lair pool.

I believe my original idea here was to have him in a tight space to show off his powers better, but I forgot to consider power scaling, he was a bit too strong for the area relative to everything else around him.
Additionally the polymorph liquid transmutation was cool, but ultimately too dangerous unless I want to up the creature to be in heaven/hell in my opinion. Heaven is where poly liquid spam is everywhere in the base game, and I want Uno to appear elsewhere.