The Hiisi fix + Fire Lukki phase preperation
Fixed all mentions of "Hisii" to be "Hiisi"
Fixed Lava Immunity not working on mod restarts, it may only take up to a minute to kick in though to prevent it being heavy on performance
Significantly Reduced likelyhood of Hungry Orb appearing in Wand Cave. It now spawns on the 1st of every month, the concept is I don't want the hungry orb to be a common item, I want it to be something you only find every few runs, something rare. Not something you can reliably get your hands on.
Reduced mana cost of Bloody Bite from 100 to 50
Increased Divine Being's durability
Reworked some of Fire Lukki's internal script to be more compact in a single file, this is mostly in preperation to handle phase changes